gts2-common: add qemu.hw.mainkeys=0 for on screen navigation bar
This commit allows user to disable the taskbar.  The on screen
navigation bar has to be enabled in order for the enable/disable taskbar
overlay to show up in settings.

Enable Taskbar
https://review.lineageos.org/q/enable+taskbar

https://forum.xda-developers.com/t/rom-unofficial-12l-eas-sm-t820-sm-t825-2022-07-15-lineageos-19-1-for-galaxy-tab-s3.4418855/page-5#post-87018775

LOS team implemented this feature a month ago, but their patch doesn't
consider the old devices which has the hardware buttons.

So, it should be an option in the "Settings - Buttons" menu, but it only
shows up for us when enabling the soft navigation bar.