gts2-common: fix offline charger by switching to LineageOS from Samsung lpm
This is a WORKAROUND. It's not the correct way of doing it so it's a temporary solution.

The Samsung lpm didn't work and shows this message

gts210wifi:/system # ./vendor/bin/lpm
CANNOT LINK EXECUTABLE "./vendor/bin/lpm": library "libpixelflinger.so" not found: needed by main executable

With the Samsung lpm, if device is powered off and then a charger was plugged in,
it would show the battery icon, but no indication of charging.  If you remove
the charger, it would still show the battery icon forcing the user to manually
shutdown the device.

Let's switch to LineageOS charger.  I tried to write all the sepolicies for it, but
even after there were no denials, the LineageOS charger would show the battery symbol
with a question mark inside it.

I discovered that if the kernel was booted as permissive, then the LineageOS charger
would charge the device showing a white color circle and lightning bolt inside.

So the idea here is that when "on shutdown" is called, we put the kernel into
permissive mode by issuing (thanks to Anan Jaser)

write /sys/fs/selinux/enforce 0

I tried using the command setenforce 0, but logs show

E [c4] init: /vendor/etc/init/hw/init.universal5433.rc: 615: Invalid keyword 'setenforce'

After the write command, the logcat showed that I needed to add 2 more sepolicies in vendor_init:

allow vendor_init kernel:security setenforce;
allow vendor_init selinuxfs:file write;

With all these hacks, offline charging now works and charges to 100% when powered off.